route Helper exceptions through async writer; fix Shipper deadlock; extract HeaderBuilder
Helper#write_exception_to_log now goes through Legion::Logging.public_send
which routes through the async writer with full segment/context metadata.
Previously it called directly on the raw Logger, bypassing async and losing
all structured context (segments, method, conv_id, etc).

Shipper rewrite eliminates AB-BA deadlock:
- Removed @flush_mutex (single @mutex, grab+clear atomically)
- stop uses @running flag + join(5) instead of Thread#kill
- start uses @start_mutex to prevent double-thread race
- Failed deliveries re-prepend batch to buffer

Extracted HeaderBuilder module to eliminate verbatim duplication of
build_exception_headers, build_exception_properties, append_identity_headers,
append_optional_header, append_legion_version_header, identity_hash,
identity_value, and redaction_enabled? between Methods and Helper.
